api_contract_review/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
openai>=1.0.0
atlassian-python-api==3.41.1
PyGithub==1.58.0
python-dotenv==1.0.0
requests==2.31.0
openapi-spec-validator==0.6.0
